+++ modules/tools/print_/mapfish3PlotService.js
@@ -588,7 +588,7 @@ const PrintModel = Tool.extend(/** @lends PrintModel.prototype */{
      */
     getVisibleLayer: function () {
         let visibleLayerList = Radio.request("Map", "getLayers").getArray().filter(layer => {
-            return layer.getVisible() === true;
+            return layer.getVisible() === true && layer.get("name") !== "markerPoint";
         });
 
         visibleLayerList = this.sortVisibleLayerListByZindex(visibleLayerList);