+++ src/modules/tools/draw/store/actions/drawInteractionOnDrawEvent.js
@@ -12,7 +12,7 @@ const errorBorder = "#E10019";
  * @returns {void}
  */
 export function drawInteractionOnDrawEvent ({state, commit, dispatch, rootState, getters}, drawInteraction) {
-    const styleSettings = getters.getStyleSettings(),
+    const {styleSettings} = getters,
         interaction = state["drawInteraction" + drawInteraction],
         circleMethod = styleSettings.circleMethod,
         drawType = state.drawType,