+++ src/modules/tools/gfi/components/themes/default/tests/unit/components/Default.spec.js
@@ -173,5 +173,36 @@ describe("src/modules/tools/gfi/components/themes/default/components/Default.vue
         expect(wrapper2.find(".gfi-theme-images > div:nth-child(2) > a > img").attributes().src).equals("https://abc.jpeg");
     });
 
+    it("should show no attribute is available message if getMappedProperties is empty", () => {
+        const wrapper1 = shallowMount(DefaultTheme, {
+            propsData: {
+                feature: {
+                    getProperties: () => {
+                        return {};
+                    },
+                    getMappedProperties: () => {
+                        return {};
+                    },
+                    getTheme: () => {
+                        return {
+                            name: "images",
+                            params: {
+                                imageLink: "abc"
+                            }
+                        };
+                    },
+                    getGfiUrl: () => "",
+                    getMimeType: () => "text/xml"
+                }
+            },
+            localVue,
+            mocks: {
+                $t: (msg) => msg
+            }
+        });
+
+        expect(wrapper1.find("td").text()).equals("modules.tools.gfi.themes.default.noAttributeAvailable");
+    });
+
 
 });