+++ test/end2end/library/scripts.js
@@ -123,6 +123,47 @@ function isLayerVisible () {
         ? layer.getVisible() && layer.getOpacity() === Number(expectedOpacity)
         : false;
 }
+/**
+ * Returns true, if mapMarker with name "markerPoint" is visible.
+ * @returns {boolean} true, if mapMarker is visible
+ */
+function isMarkerPointVisible () {
+    const map = Backbone.Radio.request("Map", "getMap");
+    let layer = false;
+
+    map.getLayers().forEach(l => {
+        if (l.get("name") === "markerPoint") {
+            layer = l;
+        }
+    });
+    if (layer) {
+        return layer.getVisible();
+    }
+    return false;
+}
+/**
+ * Returns the coordinates of the mapMarker with name "markerPoint" or null, if no marker exists.
+ * @returns {Array} containing the coordinates
+ */
+function getMarkerPointCoord () {
+    const map = Backbone.Radio.request("Map", "getMap");
+    let layer = false;
+
+    map.getLayers().forEach(l => {
+        if (l.get("name") === "markerPoint") {
+            layer = l;
+        }
+    });
+    if (layer) {
+        const feature = layer.getSource().getFeatures()[0];
+
+        if (feature) {
+            return feature.getGeometry().getCoordinates();
+        }
+    }
+    return null;
+}
+
 
 /**
  * @param {string} name name of layer to check features of
@@ -479,8 +520,10 @@ module.exports = {
     areAllLayersHidden,
     areRegExpsInMeasureLayer,
     areAllFeaturesOfLayerVisible,
+    getMarkerPointCoord,
     isFullscreen,
     isLayerVisible,
+    isMarkerPointVisible,
     imageLoaded,
     isInitalLoadingFinished,
     isObModeOn,